pyproject.toml
[tool.poetry]
name = "cumplo-scraper"
version = "0.1.0"
description = "Cumplo scraper"
authors = ["Cristobal Sfeir <cnsfeir@uc.cl>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.1"
arrow = "^1.2.3"
pydantic = "^1.10.7"
python-dotenv = "^0.21.0"
Babel = "^2.11.0"
bs4 = "^0.0.1"
aiohttp = "^3.8.3"
asyncio = "^3.4.3"
firebase-admin = "^6.0.1"
retry = "^0.9.2"
lxml = "^4.9.2"
google-cloud-logging = "^3.5.0"
starlette = "^0.27.0"
fastapi = "^0.95.2"

[tool.poetry.dev-dependencies]
pylint = "^2.17.2"
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
